updated product controller
